% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auprc.R
\name{auprc}
\alias{auprc}
\title{Calculates AUPRC measure}
\usage{
auprc(prob, y_truth, positive_value)
}
\arguments{
\item{prob}{\code{numeric} vector of probabilities of belonging to class \code{positive_value}}

\item{y_truth}{vector of "true" values of response vector}

\item{positive_value}{which class from \code{y_truth} should be treated as positive}
}
\value{
Value of area under precision-recall curve
}
\description{
Calculates AUPRC measure
}
