% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auprc.R
\name{precision_recall_curve}
\alias{precision_recall_curve}
\title{Draws AUPRC curve}
\usage{
precision_recall_curve(prob, y_truth, positive_value)
}
\arguments{
\item{prob}{\code{numeric} vector of probabilities of belonging to class \code{positive_value}}

\item{y_truth}{vector of "true" values of response vector}

\item{positive_value}{which class from \code{y_truth} should be treated as positive}
}
\value{
plot of precision-recall curve
}
\description{
Draws AUPRC curve
}
