% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auprc.R
\name{auprc}
\alias{auprc}
\title{Calculates AUPRC measure}
\usage{
auprc(prob, y_truth, positive_value)
}
\arguments{
\item{prob}{\verb{[numeric]}\cr
A vector of predicted probabilities of belonging to class
\code{positive_value}.}

\item{y_truth}{\verb{[any]}\cr
A vector of "true" labels for predicted observations. Each element
should be either \code{positive_value} or some other value, treated as
(not specified) negative value.}

\item{positive_value}{\verb{[any(1)]}\cr
Which of the two distinct values from \code{y_truth} should be treated
as positive.}
}
\value{
Value of area under precision-recall curve.
}
\description{
Calculates AUPRC measure
}
